% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizeDF.supervised.R
\name{discretizeDF.supervised}
\alias{discretizeDF.supervised}
\alias{discretize}
\title{Supervised Methods to Convert Continuous Variables into Categorical
Variables}
\usage{
discretizeDF.supervised(formula, data, method = "mdlp", dig.lab = 3, ...)
}
\arguments{
\item{formula}{a formula object to specify the class variable for supervised
discretization and the predictors to be discretized in the form
\code{class ~ .} or \code{class ~ predictor1 + predictor2}.}

\item{data}{a data.frame containing continuous variables to be discretized}

\item{method}{Discretization method. Available methods are \code{"mdlp"},
\code{"caim"}, \code{"cacc"}, \code{"ameva"}, \code{"chi2"},
\code{"chimerge"}, \code{"extendedchi2"}, and \code{"modchi2"}.}

\item{dig.lab}{integer; number of digits used to create labels.}

\item{\dots}{Additional parameters are passed on to the implementation of
the chosen discretization method.}
}
\value{
\code{discretizeDF()} returns a discretized data.frame. Discretized
columns have an attribute \code{"discretized:breaks"} indicating the
breaks used and \code{"discretized:method"} giving the method used.
}
\description{
This function implements several supervised methods to convert continuous
variables into categorical variables (factors) suitable for association
rule mining and building associative classifiers. A whole data.frame is
discretized (i.e., all numeric columns are discretized).
}
\details{
\code{discretizeDF.supervised()} only implements supervised discretization.
See \code{\link[arules:discretizeDF]{arules::discretizeDF()}} in package \pkg{arules} for unsupervised
discretization.
}
\examples{
data("iris")
summary(iris)

# supervised discretization using Species
iris.disc <- discretizeDF.supervised(Species ~ ., iris)
summary(iris.disc)

attributes(iris.disc$Sepal.Length)

# discretize the first few instances of iris using the same breaks as iris.disc
discretizeDF(head(iris), methods = iris.disc)

# only discretize predictors Sepal.Length and Petal.Length
iris.disc2 <- discretizeDF.supervised(Species ~ Sepal.Length + Petal.Length, iris)
head(iris.disc2)
}
\seealso{
Unsupervised discretization from \pkg{arules}:
\code{\link[arules:discretize]{arules::discretize()}}, \code{\link[arules:discretizeDF]{arules::discretizeDF()}}.

Details about the available supervised discretization methods from
\pkg{discretization}:
\link[discretization:mdlp]{discretization::mdlp},
\link[discretization:caim]{discretization::caim},
\link[discretization:cacc]{discretization::cacc},
\link[discretization:ameva]{discretization::ameva},
\link[discretization:chi2]{discretization::chi2},
\link[discretization:chiM]{discretization::chiM},
\link[discretization:extendChi2]{discretization::extendChi2},
\link[discretization:modChi2]{discretization::modChi2}.

Other preparation:
\code{\link[=mineCARs]{mineCARs()}},
\code{\link[=prepareTransactions]{prepareTransactions()}},
\code{\link[=transactions2DF]{transactions2DF()}}
}
\author{
Michael Hahsler
}
\concept{preparation}
\keyword{manip}
