% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAdata.R
\docType{data}
\name{Lymphography}
\alias{Lymphography}
\title{The Lymphography Domain Data Set (UCI)}
\format{
A data frame with 147 observations on the following 19 variables.
 \describe{
     \item{\code{class}}{a factor with levels \code{normalfind} \code{metastases} \code{malignlymph} \code{fibrosis}}
     \item{\code{lymphatics}}{a factor with levels \code{normal} \code{arched} \code{deformed} \code{displaced}}
     \item{\code{blockofaffere}}{a factor with levels \code{no} \code{yes}}
     \item{\code{bloflymphc}}{a factor with levels \code{no} \code{yes}}
     \item{\code{bloflymphs}}{a factor with levels \code{no} \code{yes}}
     \item{\code{bypass}}{a factor with levels \code{no} \code{yes}}
     \item{\code{extravasates}}{a factor with levels \code{no} \code{yes}}
     \item{\code{regenerationof}}{a factor with levels \code{no} \code{yes}}
     \item{\code{earlyuptakein}}{a factor with levels \code{no} \code{yes}}
     \item{\code{lymnodesdimin}}{a factor with levels \code{0} \code{1} \code{2} \code{3}}
     \item{\code{lymnodesenlar}}{a factor with levels \code{1} \code{2} \code{3} \code{4}}
     \item{\code{changesinlym}}{a factor with levels \code{bean} \code{oval} \code{round}}
     \item{\code{defectinnode}}{a factor with levels \code{no} \code{lacunar} \code{lacmarginal} \code{laccentral}}
     \item{\code{changesinnode}}{a factor with levels \code{no} \code{lacunar} \code{lacmargin} \code{laccentral}}
     \item{\code{changesinstru}}{a factor with levels \code{no} \code{grainy} \code{droplike} \code{coarse} \code{diluted} \code{reticular} \code{stripped} \code{faint}}
     \item{\code{specialforms}}{a factor with levels \code{no} \code{chalices} \code{vesicles}}
     \item{\code{dislocationof}}{a factor with levels \code{no} \code{yes}}
     \item{\code{exclusionofno}}{a factor with levels \code{no} \code{yes}}
     \item{\code{noofnodesin}}{a factor with levels \code{0-9} \code{10-19} \code{20-29} \code{30-39} \code{40-49} \code{50-59} \code{60-69} \code{>=70}}
     }
}
\source{
The data set was obtained from the UCI Machine Learning Repository
at \url{http://archive.ics.uci.edu/ml/datasets/Lymphography}.
}
\description{
This is lymphography domain obtained from the University Medical Centre,
Institute of Oncology, Ljubljana, Yugoslavia. It was repeatedly used in the
machine learning literature.
}
\examples{

data("Lymphography")

summary(Lymphography)

}
\references{
This lymphography domain was obtained from the University
Medical Centre, Institute of Oncology, Ljubljana, Yugoslavia. Thanks go to
M. Zwitter and M. Soklic for providing the data. Please include this
citation if you plan to use this database.
}
\keyword{datasets}
